[DC-1323] Move sandbox.py to utils/ directory
