+++ data_steward/cdr_cleaner/cleaning_rules/remove_invalid_procedure_source_records.py
@@ -17,9 +17,9 @@ import logging
 
 # Project imports
 import constants.cdr_cleaner.clean_cdr as cdr_consts
-from constants.cdr_cleaner import clean_cdr as clean_consts
 from constants import bq_utils as bq_consts
-from sandbox import get_sandbox_dataset_id
+from constants.cdr_cleaner import clean_cdr as clean_consts
+from utils.sandbox import get_sandbox_dataset_id
 
 LOGGER = logging.getLogger(__name__)
