+++ data_steward/cdr_cleaner/manual_cleaning_rules/update_questiona_answers_not_mapped_to_omop.py
@@ -14,7 +14,6 @@ import logging
 
 import bq_utils
 import constants.cdr_cleaner.clean_cdr as cdr_consts
-import sandbox
 
 LOGGER = logging.getLogger(__name__)
